Merged PR 936: Add code coverage and version to flutter
This PR makes flutter pull version 3.22.3 for agent installs. This is so that some issues in the current version of flutter don't cause our tests to fail. This step in the pipeline will need to change when we determine what version to keep dependencies on.

This PR also adds a coverage report to the pipeline.

https://dev.azure.com/software-engineering-studio/mHealth%20Education%20App/_build/results?buildId=2129&view=codecoverage-tab

I did **not** implement checking that code coverage didn't decrease. This is just publishing current code coverage.