Merged PR 940: Fix Windows Path for images and alter storage test order
This PR fixes ~~two~~ four items:

## 1. Some storage_provider tests started failing between Sunday 10/06 and Wednesday 10/09. I am unsure why, but it happened for me both locally and on the pipeline

The tests in local_storage_provider_test.dart that focus on Multiple files care about the order in which the files are reversed

When the test was created, the order was to look at the list in reverse. So testdata1 is at position 1 while testdata2 is at position 0.
![image (2).png](https://dev.azure.com/software-engineering-studio/34e8b4ec-91d0-44c1-855a-379b956d8bd1/_apis/git/repositories/3c2bc596-dc1e-4eeb-81bc-db0c8e4c75c5/pullRequests/940/attachments/image%20%282%29.png)

That no longer works and the results are now in the order shown below.

![image.png](https://dev.azure.com/software-engineering-studio/34e8b4ec-91d0-44c1-855a-379b956d8bd1/_apis/git/repositories/3c2bc596-dc1e-4eeb-81bc-db0c8e4c75c5/pullRequests/940/attachments/image.png)

So I just switched the order in the tests. I looked at the commit history for the packages that that module relies on and I couldn't find anything relevant that would have changed. So I am somewhat stumped as to the actual root cause.

Fortunately, only the tests take this order into account.

## 2. The image extraction process would fail when run on windows. I updated it to use the path library more liberally to be system agnostic.

Old state on Windows:
The below "Bad state: No element" errors would appear when trying to parse pptxs on Windows.
![image (3).png](https://dev.azure.com/software-engineering-studio/34e8b4ec-91d0-44c1-855a-379b956d8bd1/_apis/git/repositories/3c2bc596-dc1e-4eeb-81bc-db0c8e4c75c5/pullRequests/940/attachments/image%20%283%29.png)

Luna file now has images on Windows:
![image (4).png](https://dev.azure.com/software-engineering-studio/34e8b4ec-91d0-44c1-855a-379b956d8bd1/_apis/git/repositories/3c2bc596-dc1e-4eeb-81bc-db0c8e4c75c5/pullRequests/940/attachments/image%20%284%29.png)

##3. I also added teardown and setup steps to the image extractor tests so test files were not left around.

##4. Restored MACOS and windows files deleted during pptx folder move in [PR393](https://dev.azure.com/software-engineering-studio/mHealth%20Education%20App/_git/Luna_mHealth_Mobile/pullrequest/939)