Merged PR 967: Add Code Coverage Checks and Nightly build
This PR add a step to the lint/test pipeline to check if code coverage decreases. The pipeline already published code coverage results. Now it looks at those publish results and checks if code coverage decreases.

This PR also adds a schedule for the lint/test pipeline to run at 3:00AM every night on main.

I also added a copy of the pipeline.yml as a playground file for further development work.

If code coverage does decrease, there will be a warning, but developers can still merge their branches. I decided to make this a warning instead of a stop error because I could foresee cases where main got updated to have increased code coverage while a developer was working on a separate item. Obviously, they could just rebase their branch, but I thought just throwing a warning was a better solution

Here is an example, where code coverage decreased by .29%. The orange exclamation at the top indicates a warning instead of a total stop.

![image (2).png](https://dev.azure.com/software-engineering-studio/34e8b4ec-91d0-44c1-855a-379b956d8bd1/_apis/git/repositories/3c2bc596-dc1e-4eeb-81bc-db0c8e4c75c5/pullRequests/967/attachments/image%20%282%29.png)
You can see that the calculation is accurate

This branch
![image (3).png](https://dev.azure.com/software-engineering-studio/34e8b4ec-91d0-44c1-855a-379b956d8bd1/_apis/git/repositories/3c2bc596-dc1e-4eeb-81bc-db0c8e4c75c5/pullRequests/967/attachments/image%20%283%29.png)

Main
![image (4).png](https://dev.azure.com/software-engineering-studio/34e8b4ec-91d0-44c1-855a-379b956d8bd1/_apis/git/repositories/3c2bc596-dc1e-4eeb-81bc-db0c8e4c75c5/pullRequests/967/attachments/image%20%284%29.png)

The large number of commits is due to the fact I need to commit to test the pipeline. I'll do a squash commit when merging so that git history isn't polluted.

Related work items: #10985