Merged PR 1220: Create luna authoring run template
# Context and Change Summary
This change creates a template for pipeliens to run the luna_authoring system and confirm that a luna file is made.

Currently, the only entry into luna is using flutter run, so that is what this template does. To do so it sets the user profiles as variables depending on the OS and starts a process for flutter run via PowerShell. It then times out for a few minutes to deal with the fact that flutter run expects user input at some point.

As usually, the playground pipeline changes should be ignored.

Windows good test:

![image (2).png](https://dev.azure.com/software-engineering-studio/34e8b4ec-91d0-44c1-855a-379b956d8bd1/_apis/git/repositories/3c2bc596-dc1e-4eeb-81bc-db0c8e4c75c5/pullRequests/1220/attachments/image%20%282%29.png)

https://dev.azure.com/software-engineering-studio/mHealth%20Education%20App/_build/results?buildId=3151&view=results

MacOS good test:

![image.png](https://dev.azure.com/software-engineering-studio/34e8b4ec-91d0-44c1-855a-379b956d8bd1/_apis/git/repositories/3c2bc596-dc1e-4eeb-81bc-db0c8e4c75c5/pullRequests/1220/attachments/image.png)

https://dev.azure.com/software-engineering-studio/mHealth%20Education%20App/_build/results?buildId=3150&view=results

It doesn't work on linux. I made a bug for that.

# Notes

# Checklist

## Before submitting this PR, please make sure all boxes are checked
The full checklist is available at [Luna mHealth Code Checklist](https://docs.google.com/document/d/1uwLg5870LtvyxgchD7Cz7HWUYzQWyyVghyBRTG7YKMA/edit?usp=sharing)

### Pull Request
- [X] Right-sized PR?
- [X] Errors handled with exceptions?
- [X] Descriptive names?
- [X] Single responsibility?
- [X] Simple logic?
- [X] Follows SOLID principles?
- [X] Follows KISS principle?
- [X] Simplest algorithm?
- [X] Simplest data structure?
- [X] Focused PR?
- [X] Comments explain “why” or complex “how”?
- [X] Class documentation added?
- [X] Method comments define contract?
- [X] No sensitive info?

### Understandability
- [X] Understandable code?
- [X] Short methods?
- [X] Low cyclomatic complexity?
- [X] No magic literals?
- [X] New branch per story/bug?
- [X] Telemetry added?

### Maintainability
- [X] Follows DRY?
- [X] No duplicate lines?
- [X] No redundant patterns?

### Test Coverage
- [X] Unit test each path?
- [X] Single test case per unit test?
- [X] Expected exceptions tested?
- [X] No dev environment updates?

### Correctness
  - [X] Inputs validated?
  - [X] Handles incorrect inputs?
  - [X] Correct error responses?
  - [X] Logical endpoints/verbs?
  - [X] Handled edge cases?

Related work items: #12332