Merged PR 1260: Compress Nightly Build to 2 Jobs
# Context and Code Change
The nightly build takes between 50 minutes and 1 hour.

Most of the code changes is a reordering of the jobs. I did have to duplicate some code due to getting rid of the strategy matrix because that spawns a new job in the matrix.

This compressed build still takes 40 minutes.

![image.png](https://dev.azure.com/software-engineering-studio/34e8b4ec-91d0-44c1-855a-379b956d8bd1/_apis/git/repositories/3c2bc596-dc1e-4eeb-81bc-db0c8e4c75c5/pullRequests/1260/attachments/image.png)

https://dev.azure.com/software-engineering-studio/mHealth%20Education%20App/_build/results?buildId=3300&view=results

However that is better than the recent 1 hour runs

![image (2).png](https://dev.azure.com/software-engineering-studio/34e8b4ec-91d0-44c1-855a-379b956d8bd1/_apis/git/repositories/3c2bc596-dc1e-4eeb-81bc-db0c8e4c75c5/pullRequests/1260/attachments/image%20%282%29.png)

![image (3).png](https://dev.azure.com/software-engineering-studio/34e8b4ec-91d0-44c1-855a-379b956d8bd1/_apis/git/repositories/3c2bc596-dc1e-4eeb-81bc-db0c8e4c75c5/pullRequests/1260/attachments/image%20%283%29.png)