Deleting Itme and Item_Type classes
In this commit I am removing the Item and Item_Type classes from the codebase.
This is part of the refactoring process to improve the code structure and
maintainability. The Item class was previously used to represent items in the
application, but it has been determined that it is no longer necessary.
The Item_Type class was also removed as it was closely tied to the Item class.

Also modified the following files to remove references to the Item and Item_Type classes:
- component.dart
- image_component.dart
- module.dart
- page.dart
- module_test.dart
- page_builder_service.dart