Merged PR 1327: Add load module helper function to patrol tests
# Context
The current load module test in the end-to-end tests for the mobile app has functionality that is useful for testing several different luna modules.

This PR extracts the portion of the current test that loads a Luna module to a dedicated function in the test. This allows for writing more tests to test other Luna modules without having to rewrite a bunch of the same load module code.

The testing pipeline code just changed because I was using it to test this change.

Successful test:
https://dev.azure.com/software-engineering-studio/mHealth%20Education%20App/_build/results?buildId=3490&view=results

# Checklist

## Before submitting this PR, please make sure all boxes are checked
The full checklist is available at [Luna mHealth Code Checklist](https://docs.google.com/document/d/1uwLg5870LtvyxgchD7Cz7HWUYzQWyyVghyBRTG7YKMA/edit?usp=sharing)

### Pull Request
- [X] Right-sized PR?
- [X] Errors handled with exceptions?
- [X] Descriptive names?
- [X] Single responsibility?
- [X] Simple logic?
- [X] Follows SOLID principles?
- [X] Follows KISS principle?
- [X] Simplest algorithm?
- [X] Simplest data structure?
- [X] Focused PR?
- [X] Comments explain “why” or complex “how”?
- [X] Class documentation added?
- [X] Method comments define contract?
- [X] No sensitive info?

### Understandability
- [X] Understandable code?
- [X] Short methods?
- [X] Low cyclomatic complexity?
- [X] No magic literals?
- [X] New branch per story/bug?
- [X] Telemetry added?

### Maintainability
- [X] Follows DRY?
- [X] No duplicate lines?
- [X] No redundant patterns?

### Test Coverage
- [X] Unit test each path?
- [X] Single test case per unit test?
- [X] Expected exceptions tested?
- [X] No dev environment updates?

### Correctness
  - [X] Inputs validated?
  - [X] Handles incorrect inputs?
  - [X] Correct error responses?
  - [X] Logical endpoints/verbs?
  - [X] Handled edge cases?