Merged PR 1326: Fix vncserver startup issue
# Context
When starting the dev container, the vncserver is not starting. This was due to a logging change that I made that I spose I didn't test before my last PR.

I fixed this by removing nohup and just running vncserver in the background.

Since this just a change to the startup script, it will take affect on container restart and does not require a rebuild.