Merged PR 1335: Allow nightly build to test multiple different luna modules
# Context
Currently the e2e tests for mobile is essentially hard coded to only use the "A line.pptx" for testing. This obviously wont work as we add more and more functionality.

This change updates the pipeline ymls and e2e test file to allow for a list of files to be used in e2e integration.

# Change Summary

**testAssets**
I created a new test Assets folder for the e2e tests. This does mean some test assets are duplicated, but it makes for troubleshooting the e2e tests on the pipelines easier.

**luna_mobile_e2e_tests**
The e2e tests patrol test now will loop through a map of files in the form

"LUNA MODULE NAME.LUNA":"LUNA TITLE"

**run_luna_mobile_tests_firebase.yml**
This will now loop through a parameter of luna module names and file names. For each file, it will run Luna authoring against that file in the testAssets folder and then add it to a string of file names to give to firebase for mobile testing

Example:

      - luna_mod_name: "a_red_line_test"
        input_file_name: "A red line.pptx"
      - luna_mod_name: "a_line_test"
        input_file_name: "A line.pptx"

**pipeline_nightly_build.yml**
Delete the run authoring line since, the authoring system is run in the mobile tests for generating files.

# Notes

# Checklist

## Before submitting this PR, please make sure all boxes are checked
The full checklist is available at [Luna mHealth Code Checklist](https://docs.google.com/document/d/1uwLg5870LtvyxgchD7Cz7HWUYzQWyyVghyBRTG7YKMA/edit?usp=sharing)

### Pull Request
- [ ] Right-sized PR?
- [ ] Errors handled with exceptions?
- [X] Descriptive names?
- [ ] Single responsibility?
- [X] Simple logic?
- [ ] Follows SOLID principles?
- [X] Follows KISS principle?
- [X] Simplest algorithm?
- [ ] Simplest data structure?
- [X] Focused PR?
- [X] Comments explain “why” or complex “how”?
- [X] Class documentation added?
- [X] Method comments define contract?
- [X] No sensitive info?

### Understandability
- [ ] Understandable code?
- [X] Short methods?
- [X] Low cyclomatic complexity?
- [ ] No magic literals?
- [ ] New branch per story/bug?
- [ ] Telemetry added?

### Maintainability
- [X] Follows DRY?
- [ ] No duplicate lines?
- [X] No redundant patterns?

### Test Coverage
- [ ] Unit test each path?
- [ ] Single test case per unit test?
- [ ] Expected exceptions tested?
- [ ] No dev environment updates?

### Correctness
  - [ ] Inputs validated?
  - [ ] Handles incorrect inputs?
  - [X] Correct error responses?
  - [X] Logical endpoints/verbs?
  - [ ] Handled edge cases?