Merged PR 1349: Create Python Script to Update Files in Google Drive
# Context
Currently, there is not a good place to place .exes, .apps, and .abds generated by the nightly build. They are technically accessible in the pipelines artifacts for the most recent generated pipeline.

However, that is not an easily accessible or obvious place to store those files. Google Drive stores many files for the Luna project. I created a Releases folder for storing these artifacts further.

# Change Summary

I created a new python script to updated files in Google Drive. This script _intentionally_ updates files so that the script can't create or delete new files. I used python because it is a good scripting language and the documentation for the Google drive API was verbose for python. Additionally, there is an ADO builtin task for calling python scripts in pipelines.

I manually uploaded three zip files for the builds that will be updated by this script. The IDs for those zip files are stored in a variable group in ADO for access in the pipelines.

![image.png](https://dev.azure.com/software-engineering-studio/34e8b4ec-91d0-44c1-855a-379b956d8bd1/_apis/git/repositories/3c2bc596-dc1e-4eeb-81bc-db0c8e4c75c5/pullRequests/1349/attachments/image.png)

Below are examples of where I updated a zip file using the script.

![image (2).png](https://dev.azure.com/software-engineering-studio/34e8b4ec-91d0-44c1-855a-379b956d8bd1/_apis/git/repositories/3c2bc596-dc1e-4eeb-81bc-db0c8e4c75c5/pullRequests/1349/attachments/image%20%282%29.png)

# Checklist

## Before submitting this PR, please make sure all boxes are checked
The full checklist is available at [Luna mHealth Code Checklist](https://docs.google.com/document/d/1uwLg5870LtvyxgchD7Cz7HWUYzQWyyVghyBRTG7YKMA/edit?usp=sharing)

### Pull Request
- [X] Right-sized PR?
- [ ] Errors handled with exceptions?
- [X] Descriptive names?
- [X] Single responsibility?
- [X] Simple logic?
- [X] Follows SOLID principles?
- [X] Follows KISS principle?
- [X] Simplest algorithm?
- [ ] Simplest data structure?
- [ ] Focused PR?
- [X] Comments explain “why” or complex “how”?
- [ ] Class documentation added?
- [X] Method comments define contract?
- [X] No sensitive info?

### Understandability
- [X] Understandable code?
- [X] Short methods?
- [X] Low cyclomatic complexity?
- [X] No magic literals?
- [X] New branch per story/bug?
- [ ] Telemetry added?

### Maintainability
- [X] Follows DRY?
- [X] No duplicate lines?
- [X] No redundant patterns?

### Test Coverage
- [ ] Unit test each path?
- [ ] Single test case per unit test?
- [ ] Expected exceptions tested?
- [ ] No dev environment updates?

### Correctness
  - [ ] Inputs validated?
  - [ ] Handles incorrect inputs?
  - [ ] Correct error responses?
  - [ ] Logical endpoints/verbs?
  - [ ] Handled edge cases?

Related work items: #12424