Merged PR 1369: Split out Patrol App Buildout
# Context
With the upgrade to Java 17 to support the newer version of flutter, the JVM keeps crashing during mobile builds. For the actual mobile build, we switched it to linux. This change does the same thing with the patrol build. It moved building the patrol build, for integration testing, from the Windows job to a separate Linux job.

Additionally, the gcloud web call wouldn't error out the entire task. It just spits out the error and the task assumes it was successful. So I updated the Powershell used to call Firebase to fail if the call to gcloud fails.

The build.gradle was updated to specify java 17. This is only needed for patrol testing since that file is not used in the actual build of the mobile app.

Full test:
https://dev.azure.com/software-engineering-studio/mHealth%20Education%20App/_build/results?buildId=3693&view=results

Purposefully failed test:

Task now fails when gcloud to Firebase call fails

https://dev.azure.com/software-engineering-studio/mHealth%20Education%20App/_build/results?buildId=3692&view=results

# Notes

# Checklist

## Before submitting this PR, please make sure all boxes are checked
The full checklist is available at [Luna mHealth Code Checklist](https://docs.google.com/document/d/1uwLg5870LtvyxgchD7Cz7HWUYzQWyyVghyBRTG7YKMA/edit?usp=sharing)

### Pull Request
- [X] Right-sized PR?
- [X] Errors handled with exceptions?
- [X] Descriptive names?
- [X] Single responsibility?
- [X] Simple logic?
- [ ] Follows SOLID principles?
- [ ] Follows KISS principle?
- [ ] Simplest algorithm?
- [ ] Simplest data structure?
- [X] Focused PR?
- [ ] Comments explain “why” or complex “how”?
- [ ] Class documentation added?
- [ ] Method comments define contract?
- [X] No sensitive info?

### Understandability
- [X] Understandable code?
- [ ] Short methods?
- [ ] Low cyclomatic complexity?
- [ ] No magic literals?
- [ ] New branch per story/bug?
- [ ] Telemetry added?

### Maintainability
- [X] Follows DRY?
- [X] No duplicate lines?
- [X] No redundant patterns?

### Test Coverage
- [ ] Unit test each path?
- [ ] Single test case per unit test?
- [ ] Expected exceptions tested?
- [ ] No dev environment updates?

### Correctness
  - [ ] Inputs validated?
  - [ ] Handles incorrect inputs?
  - [ ] Correct error responses?
  - [ ] Logical endpoints/verbs?
  - [ ] Handled edge cases?