Merged PR 1382: Create Git Tag Version Tempalte
# Context
(The changes to the playground yml are used for testing pipelines)

Currently the only place that the version number of Luna lives is in the pipeline variable library. That number is then passed into the build for each build of Luna.

This change creates a template to start moving to using Git tags for versioning.

Why use git tags?

1. The tags section of ADO shows each tag.
1a) Currently these are all nonsense tags as I was doing my testing. Once we actually start using tags, I will remediate this list.

![image.png](https://dev.azure.com/software-engineering-studio/34e8b4ec-91d0-44c1-855a-379b956d8bd1/_apis/git/repositories/3c2bc596-dc1e-4eeb-81bc-db0c8e4c75c5/pullRequests/1382/attachments/image.png)

2. When browsing the repo, you can just select tags

![image (2).png](https://dev.azure.com/software-engineering-studio/34e8b4ec-91d0-44c1-855a-379b956d8bd1/_apis/git/repositories/3c2bc596-dc1e-4eeb-81bc-db0c8e4c75c5/pullRequests/1382/attachments/image%20%282%29.png)

3. Using tags in this manner is fairly common in industry. https://git-scm.com/book/en/v2/Git-Basics-Tagging

# Change Summary
This change doesn't change the build process. It just creates a template to use when we switch to using git tags.

This template will:

Load all tags match the format of NUMBER.NUMBER.NUMBER via a regex match.
It then sorts it to find the highest numbered version, which git has builtin functionality to do, and then creates a new tag and pushes it to the repo. Tags are not branch specific and such this push is not a commit.

If there are more than 50 tags, the lowest numbered tag is deleted. This process ignores any non-patch versions such as 1.1.0 or 0.1.0.

Pipeline working with this template:
![image (4).png](https://dev.azure.com/software-engineering-studio/34e8b4ec-91d0-44c1-855a-379b956d8bd1/_apis/git/repositories/3c2bc596-dc1e-4eeb-81bc-db0c8e4c75c5/pullRequests/1382/attachments/image%20%284%29.png)

Deletion code working:
![image (3).png](https://dev.azure.com/software-engineering-studio/34e8b4ec-91d0-44c1-855a-379b956d8bd1/_apis/git/repositories/3c2bc596-dc1e-4eeb-81bc-db0c8e4c75c5/pullRequests/1382/attachments/image%20%283%29.png)

# Notes

# Checklist

## Before submitting this PR, please make sure all boxes are checked
The full checklist is available at [Luna mHealth Code Checklist](https://docs.google.com/document/d/1uwLg5870LtvyxgchD7Cz7HWUYzQWyyVghyBRTG7YKMA/edit?usp=sharing)

### Pull Request
- [X] Right-sized PR?
- [ ] Errors handled with exceptions?
- [X] Descriptive names?
- [X] Single responsibility?
- [X] Simple logic?
- [X] Follows SOLID principles?
- [X] Follows KISS principle?
- [ ] Simplest algorithm?
- [ ] Simplest data structure?
- [X] Focused PR?
- [X] Comments explain “why” or complex “how”?
- [ ] Class documentation added?
- [ ] Method comments define contract?
- [X] No sensitive info?

### Understandability
- [X] Understandable code?
- [X] Short methods?
- [X] Low cyclomatic complexity?
- [ ] No magic li...