Merged PR 1384: Use Git Tags to Track Versions
# Context
Currently the version of our app is tracked via a variable in the pipeline library and the name of the build pipelines. That is pretty hard to follow.

This change updates the build pipeline to use the update_version template which utilizes git tags to track versions.

This change also combines the luna mobile and luna authoring pipelines to one pipeline. This one pipeline has a job for mobile and a job for authoring. It can selectively build one or the authoring based on parameters. By default, both are built.

![image (2).png](https://dev.azure.com/software-engineering-studio/34e8b4ec-91d0-44c1-855a-379b956d8bd1/_apis/git/repositories/3c2bc596-dc1e-4eeb-81bc-db0c8e4c75c5/pullRequests/1384/attachments/image%20%282%29.png)

After this change, I will delete the current tags and set the apps version to "0.0.22". This is based on the last built Luna app version.

![image.png](https://dev.azure.com/software-engineering-studio/34e8b4ec-91d0-44c1-855a-379b956d8bd1/_apis/git/repositories/3c2bc596-dc1e-4eeb-81bc-db0c8e4c75c5/pullRequests/1384/attachments/image.png)

I also made a small update to the update_version.yml template due to an issue with Powershell returning a list vs a string depending on how many tags are in the result of "git tag".

Successful build:

https://dev.azure.com/software-engineering-studio/mHealth%20Education%20App/_build/results?buildId=3783&view=results

![image (3).png](https://dev.azure.com/software-engineering-studio/34e8b4ec-91d0-44c1-855a-379b956d8bd1/_apis/git/repositories/3c2bc596-dc1e-4eeb-81bc-db0c8e4c75c5/pullRequests/1384/attachments/image%20%283%29.png)

Created in Google drive

![image (4).png](https://dev.azure.com/software-engineering-studio/34e8b4ec-91d0-44c1-855a-379b956d8bd1/_apis/git/repositories/3c2bc596-dc1e-4eeb-81bc-db0c8e4c75c5/pullRequests/1384/attachments/image%20%284%29.png)

# Notes

# Checklist

## Before submitting this PR, please make sure all boxes are checked
The full checklist is available at [Luna mHealth Code Checklist](https://docs.google.com/document/d/1uwLg5870LtvyxgchD7Cz7HWUYzQWyyVghyBRTG7YKMA/edit?usp=sharing)

### Pull Request
- [X] Right-sized PR?
- [X] Errors handled with exceptions?
- [X] Descriptive names?
- [X] Single responsibility?
- [X] Simple logic?
- [X] Follows SOLID principles?
- [X] Follows KISS principle?
- [ ] Simplest algorithm?
- [ ] Simplest data structure?
- [X] Focused PR?
- [X] Comments explain “why” or complex “how”?
- [X] Class documentation added?
- [ ] Method comments define contract?
- [X] No sensitive info?

### Understandability
- [X] Understandable code?
- [X] Short methods?
- [ ] Low cyclomatic complexity?
- [X] No magic literals?
- [X] New branch per story/bug?
- [ ] Telemetry added?

### Maintainability
- [ ] Follows DRY?
- [ ] No duplicate lines?
- [ ] No redundant patterns?

### Test Coverage
- [ ] Unit test each path?
- [ ] Single test case per unit test?
- [ ] Expected exceptions tested?
- [ ] No dev environment updates?

### Correctn...